pyproject.toml
[tool.poetry]
name = "automatic-mail-service"
version = "0.1.0"
description = "App To Handle Event Mail Automation"
authors = ["Kurniadi Ahmad Wijaya <kurniadiahmadwijaya@gmail.com>"]
readme = "README.md"
packages = [{include = "automatic_mail_service"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
sqlalchemy = "^2.0.19"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.5"
marshmallow = "^3.20.1"
pymysql = "^1.1.0"
flask-mail = "^0.9.1"
flask-restful = "^0.3.10"
cryptography = "^41.0.2"
gunicorn = "^21.2.0"
rq = "^1.15.1"
rq-scheduler = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
